Slowed down bag playback in test. Hopefully should help prevent test-startup race condition
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@15343 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 978127ab0d7276e1ebd6baa9e4549db7cf9b861a

+++ drivers/robot/pr2/point_cloud_assembler/test/test_assembler.cpp
@@ -135,7 +135,7 @@ TEST_F(TestAssembler, test)
 
   usleep(1e6) ;
 
-  ASSERT_EQ(cloud_counter_, (unsigned int)5) ;
+  ASSERT_EQ(cloud_counter_, 5) ;
 
   unsigned int cloud_size ;
   cloud_mutex_.lock() ;
